+++ mobile/src/main/java/com/boswelja/devicemanager/managespace/ui/sheets/ResetAppBottomSheet.kt
@@ -1,20 +1,21 @@
-package com.boswelja.devicemanager.managespace.ui
+package com.boswelja.devicemanager.managespace.ui.sheets
 
+import android.app.ActivityManager
+import android.content.Context
 import androidx.core.content.ContextCompat
 import com.boswelja.devicemanager.R
 
 /**
- * A [BaseResetBottomSheet] to handle resetting all app settings. This does not include extension
- * settings.
+ * A [BaseResetBottomSheet] to handle resetting all of Wearable Extensions.
  */
-class ResetAppSettingsBottomSheet : BaseResetBottomSheet() {
+class ResetAppBottomSheet : BaseResetBottomSheet() {
 
     private var successful = false
 
     override fun onCreateConfirmation(showProgress: () -> Unit): ConfirmationFragment {
         return ConfirmationFragment(
-            getString(R.string.reset_settings_title),
-            getString(R.string.reset_settings_desc),
+            getString(R.string.reset_app_title),
+            getString(R.string.reset_app_warning),
             getString(R.string.dialog_button_reset),
             ContextCompat.getDrawable(requireContext(), R.drawable.ic_chevron_right)!!,
             showProgress
@@ -23,30 +24,33 @@ class ResetAppSettingsBottomSheet : BaseResetBottomSheet() {
 
     override fun onCreateProgress(): ProgressFragment {
         return ProgressFragment(
-            getString(R.string.reset_settings_resetting),
+            getString(R.string.reset_app_resetting),
             getString(R.string.please_wait)
         )
     }
 
     override fun onCreateResult(dismissSheet: () -> Unit): ConfirmationFragment {
+        val activityManager = requireContext()
+            .getSystemService(Context.ACTIVITY_SERVICE) as ActivityManager
+
         val title = if (successful)
-            getString(R.string.reset_settings_success)
+            getString(R.string.reset_app_success)
         else
-            getString(R.string.reset_settings_failed)
+            getString(R.string.reset_app_failed)
 
         return ConfirmationFragment(
             title,
-            "",
-            getString(R.string.button_done),
+            getString(R.string.reset_app_restarting_note),
+            getString(R.string.button_finish),
             ContextCompat.getDrawable(requireContext(), R.drawable.ic_check)!!
         ) {
+            if (successful) activityManager.clearApplicationUserData()
             dismissSheet()
-            activity?.recreate()
         }
     }
 
     override fun onStartWork() {
-        viewModel.resetAppSettings(
+        viewModel.resetApp(
             { setProgress(it) },
             {
                 successful = it
